excel-to-xmind/requirements.txt
pandas
xmind-sdk